feat : landingpage 구현, 컨테이너 w,h 반응형 조정
