+++ package.json
@@ -9,6 +9,8 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@dotlottie/player-component": "^2.7.12",
+    "autoprefixer": "^10.4.19",
     "axios": "^1.6.8",
     "next": "14.2.3",
     "prettier": "^3.2.5",
@@ -22,8 +24,8 @@
     "@types/react-dom": "^18",
     "eslint": "^8",
     "eslint-config-next": "14.2.3",
-    "postcss": "^8",
-    "tailwindcss": "^3.4.1",
+    "postcss": "^8.4.38",
+    "tailwindcss": "^3.4.3",
     "typescript": "^5"
   }
 }
