+++ src/app/layout.tsx
@@ -1,3 +1,4 @@
+
 import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
 import localFont from 'next/font/local';
@@ -30,12 +31,10 @@ export default function RootLayout({
     <html lang="ko" className={`${pretendard.variable}`}>
       <body className={pretendard.className}>
         {/* Full */}
-        <div className="w-full h-screen flex justify-center items-start overflow-hidden g-gradient-to-r from-gray-300 via-gray-400 to-gray-600">
+        <div className="w-full min-h-screen flex justify-center items-start overflow-hidden gradient-bg">
           {/* Mobile */}
-          <div className="md:w-[375px] w-full h-screen  bg-black overflow-hidden relative md:rounded-xl overflow-y-visible scrollbar-hide">
-            <Header />
+          <div className="w-full md:max-w-[375px] h-screen bg-black overflow-hidden relative md:rounded-xl overflow-y-visible scrollbar-hide">
             {children}
-            <Footer />
           </div>
         </div>
       </body>
